requirements.txt
# Core dependencies
tensorflow>=2.13.0,<2.16.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional but recommended
pandas>=2.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0
